# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Product Perception and Keyword Intelligence API",
    "version": "0.1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Health Health Get"
                }
              }
            }
          }
        }
      }
    },
    "/ready": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Ready",
        "operationId": "ready_ready_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Ready Ready Get"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/perception-runs": {
      "post": {
        "tags": [
          "perception"
        ],
        "summary": "Create Perception Run",
        "operationId": "create_perception_run_api_v1_perception_runs_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_perception_run_api_v1_perception_runs_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PerceptionRun"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/perception-runs/{run_id}": {
      "get": {
        "tags": [
          "perception"
        ],
        "summary": "Get Perception Run",
        "operationId": "get_perception_run_api_v1_perception_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PerceptionRun"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_create_perception_run_api_v1_perception_runs_post": {
        "properties": {
          "images": {
            "items": {
              "type": "string",
              "contentMediaType": "application/octet-stream"
            },
            "type": "array",
            "title": "Images",
            "description": "One to five JPG, PNG, or WebP images"
          },
          "description": {
            "type": "string",
            "minLength": 1,
            "title": "Description"
          },
          "brand": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand"
          },
          "market": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "category_hint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Hint"
          },
          "target_audience_hint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Audience Hint"
          },
          "include_debug": {
            "type": "boolean",
            "title": "Include Debug",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "images",
          "description"
        ],
        "title": "Body_create_perception_run_api_v1_perception_runs_post"
      },
      "ClaimFlag": {
        "properties": {
          "claim": {
            "type": "string",
            "minLength": 1,
            "title": "Claim"
          },
          "reason": {
            "type": "string",
            "minLength": 1,
            "title": "Reason"
          },
          "severity": {
            "type": "string",
            "pattern": "^(info|warning|blocker)$",
            "title": "Severity",
            "default": "warning"
          },
          "evidence_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "claim",
          "reason"
        ],
        "title": "ClaimFlag"
      },
      "EnrichmentMetrics": {
        "properties": {
          "average_monthly_searches": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Monthly Searches"
          },
          "competition_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Competition Level"
          },
          "cpc_low": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cpc Low"
          },
          "cpc_high": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cpc High"
          },
          "trend": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trend"
          },
          "source_confidence": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Confidence"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "EnrichmentMetrics"
      },
      "EvidenceLinkedText": {
        "properties": {
          "value": {
            "type": "string",
            "minLength": 1,
            "title": "Value"
          },
          "evidence_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Evidence Ids"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "value",
          "evidence_ids",
          "confidence"
        ],
        "title": "EvidenceLinkedText"
      },
      "EvidenceRecord": {
        "properties": {
          "id": {
            "type": "string",
            "minLength": 3,
            "title": "Id"
          },
          "source": {
            "$ref": "#/components/schemas/EvidenceSource"
          },
          "source_reference": {
            "type": "string",
            "minLength": 1,
            "title": "Source Reference"
          },
          "observation": {
            "type": "string",
            "minLength": 1,
            "title": "Observation"
          },
          "quote": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Quote"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "source",
          "source_reference",
          "observation",
          "confidence"
        ],
        "title": "EvidenceRecord"
      },
      "EvidenceSource": {
        "type": "string",
        "enum": [
          "user_description",
          "user_metadata",
          "image_observation",
          "model_inference",
          "keyword_provider"
        ],
        "title": "EvidenceSource"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ImageInput": {
        "properties": {
          "index": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Index"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "mime_type": {
            "type": "string",
            "title": "Mime Type"
          },
          "content_hash": {
            "type": "string",
            "title": "Content Hash"
          },
          "byte_size": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Byte Size"
          },
          "width": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Width"
          },
          "height": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Height"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "index",
          "filename",
          "mime_type",
          "content_hash",
          "byte_size",
          "width",
          "height"
        ],
        "title": "ImageInput"
      },
      "KeywordCandidate": {
        "properties": {
          "text": {
            "type": "string",
            "minLength": 1,
            "title": "Text"
          },
          "normalized_text": {
            "type": "string",
            "minLength": 1,
            "title": "Normalized Text"
          },
          "intent": {
            "$ref": "#/components/schemas/KeywordIntent"
          },
          "category": {
            "$ref": "#/components/schemas/KeywordCategory"
          },
          "rationale": {
            "type": "string",
            "minLength": 1,
            "title": "Rationale"
          },
          "source": {
            "type": "string",
            "minLength": 1,
            "title": "Source"
          },
          "evidence_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Evidence Ids"
          },
          "relevance_score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Relevance Score"
          },
          "confidence_score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence Score"
          },
          "score_components": {
            "$ref": "#/components/schemas/ScoreComponents"
          },
          "risk_flags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Risk Flags"
          },
          "enrichment": {
            "$ref": "#/components/schemas/EnrichmentMetrics"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text",
          "normalized_text",
          "intent",
          "category",
          "rationale",
          "source",
          "evidence_ids",
          "relevance_score",
          "confidence_score",
          "score_components"
        ],
        "title": "KeywordCandidate"
      },
      "KeywordCategory": {
        "type": "string",
        "enum": [
          "product",
          "feature",
          "benefit",
          "problem_solution",
          "use_case",
          "audience",
          "long_tail",
          "alternative",
          "negative",
          "content_angle"
        ],
        "title": "KeywordCategory"
      },
      "KeywordCluster": {
        "properties": {
          "id": {
            "type": "string",
            "minLength": 3,
            "title": "Id"
          },
          "theme": {
            "type": "string",
            "minLength": 1,
            "title": "Theme"
          },
          "primary_keyword": {
            "type": "string",
            "minLength": 1,
            "title": "Primary Keyword"
          },
          "member_keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Member Keywords"
          },
          "dominant_intent": {
            "$ref": "#/components/schemas/KeywordIntent"
          },
          "category": {
            "$ref": "#/components/schemas/KeywordCategory"
          },
          "aggregate_relevance": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Aggregate Relevance"
          },
          "evidence_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Evidence Ids"
          },
          "recommended_usage": {
            "type": "string",
            "minLength": 1,
            "title": "Recommended Usage"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "theme",
          "primary_keyword",
          "member_keywords",
          "dominant_intent",
          "category",
          "aggregate_relevance",
          "evidence_ids",
          "recommended_usage"
        ],
        "title": "KeywordCluster"
      },
      "KeywordIntent": {
        "type": "string",
        "enum": [
          "informational",
          "commercial",
          "transactional",
          "comparison",
          "navigational",
          "unknown"
        ],
        "title": "KeywordIntent"
      },
      "PerceptionRun": {
        "properties": {
          "schema_version": {
            "type": "string",
            "title": "Schema Version",
            "default": "2026-06-23.mvp1b.v1"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "request": {
            "$ref": "#/components/schemas/ProductAnalysisRequest"
          },
          "images": {
            "items": {
              "$ref": "#/components/schemas/ImageInput"
            },
            "type": "array",
            "title": "Images"
          },
          "product_profile": {
            "$ref": "#/components/schemas/ProductProfile"
          },
          "keyword_candidates": {
            "items": {
              "$ref": "#/components/schemas/KeywordCandidate"
            },
            "type": "array",
            "title": "Keyword Candidates"
          },
          "keyword_clusters": {
            "items": {
              "$ref": "#/components/schemas/KeywordCluster"
            },
            "type": "array",
            "title": "Keyword Clusters"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Errors"
          },
          "stage_statuses": {
            "items": {
              "$ref": "#/components/schemas/StageStatus"
            },
            "type": "array",
            "title": "Stage Statuses"
          },
          "metadata": {
            "$ref": "#/components/schemas/ProviderMetadata"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "request",
          "images",
          "product_profile",
          "keyword_candidates",
          "keyword_clusters",
          "stage_statuses",
          "metadata"
        ],
        "title": "PerceptionRun"
      },
      "ProductAnalysisRequest": {
        "properties": {
          "description": {
            "type": "string",
            "minLength": 1,
            "title": "Description"
          },
          "brand": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand"
          },
          "market": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "category_hint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Hint"
          },
          "target_audience_hint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Audience Hint"
          },
          "include_debug": {
            "type": "boolean",
            "title": "Include Debug",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "description"
        ],
        "title": "ProductAnalysisRequest"
      },
      "ProductProfile": {
        "properties": {
          "product_name": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EvidenceLinkedText"
              },
              {
                "type": "null"
              }
            ]
          },
          "brand": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EvidenceLinkedText"
              },
              {
                "type": "null"
              }
            ]
          },
          "category": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EvidenceLinkedText"
              },
              {
                "type": "null"
              }
            ]
          },
          "subcategory": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EvidenceLinkedText"
              },
              {
                "type": "null"
              }
            ]
          },
          "summary": {
            "$ref": "#/components/schemas/EvidenceLinkedText"
          },
          "visual_attributes": {
            "items": {
              "$ref": "#/components/schemas/EvidenceLinkedText"
            },
            "type": "array",
            "title": "Visual Attributes"
          },
          "observed_facts": {
            "items": {
              "$ref": "#/components/schemas/EvidenceLinkedText"
            },
            "type": "array",
            "title": "Observed Facts"
          },
          "user_provided_facts": {
            "items": {
              "$ref": "#/components/schemas/EvidenceLinkedText"
            },
            "type": "array",
            "title": "User Provided Facts"
          },
          "inferred_attributes": {
            "items": {
              "$ref": "#/components/schemas/EvidenceLinkedText"
            },
            "type": "array",
            "title": "Inferred Attributes"
          },
          "features": {
            "items": {
              "$ref": "#/components/schemas/EvidenceLinkedText"
            },
            "type": "array",
            "title": "Features"
          },
          "benefits": {
            "items": {
              "$ref": "#/components/schemas/EvidenceLinkedText"
            },
            "type": "array",
            "title": "Benefits"
          },
          "materials": {
            "items": {
              "$ref": "#/components/schemas/EvidenceLinkedText"
            },
            "type": "array",
            "title": "Materials"
          },
          "colors": {
            "items": {
              "$ref": "#/components/schemas/EvidenceLinkedText"
            },
            "type": "array",
            "title": "Colors"
          },
          "use_cases": {
            "items": {
              "$ref": "#/components/schemas/EvidenceLinkedText"
            },
            "type": "array",
            "title": "Use Cases"
          },
          "target_audiences": {
            "items": {
              "$ref": "#/components/schemas/EvidenceLinkedText"
            },
            "type": "array",
            "title": "Target Audiences"
          },
          "differentiators": {
            "items": {
              "$ref": "#/components/schemas/EvidenceLinkedText"
            },
            "type": "array",
            "title": "Differentiators"
          },
          "limitations": {
            "items": {
              "$ref": "#/components/schemas/EvidenceLinkedText"
            },
            "type": "array",
            "title": "Limitations"
          },
          "ambiguities": {
            "items": {
              "$ref": "#/components/schemas/EvidenceLinkedText"
            },
            "type": "array",
            "title": "Ambiguities"
          },
          "unknowns": {
            "items": {
              "$ref": "#/components/schemas/EvidenceLinkedText"
            },
            "type": "array",
            "title": "Unknowns"
          },
          "unsafe_or_unverified_claims": {
            "items": {
              "$ref": "#/components/schemas/ClaimFlag"
            },
            "type": "array",
            "title": "Unsafe Or Unverified Claims"
          },
          "claim_flags": {
            "items": {
              "$ref": "#/components/schemas/ClaimFlag"
            },
            "type": "array",
            "title": "Claim Flags"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceRecord"
            },
            "type": "array",
            "minItems": 1,
            "title": "Evidence"
          },
          "overall_confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Overall Confidence"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "summary",
          "evidence",
          "overall_confidence"
        ],
        "title": "ProductProfile"
      },
      "ProviderMetadata": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "latency_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Ms"
          },
          "prompt_version": {
            "type": "string",
            "title": "Prompt Version"
          },
          "usage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProviderUsage"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "provider",
          "model",
          "prompt_version"
        ],
        "title": "ProviderMetadata"
      },
      "ProviderUsage": {
        "properties": {
          "input_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Tokens"
          },
          "output_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Tokens"
          },
          "total_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Tokens"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ProviderUsage"
      },
      "ScoreComponents": {
        "properties": {
          "product_match": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Product Match"
          },
          "intent_value": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Intent Value"
          },
          "evidence_strength": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Evidence Strength"
          },
          "audience_fit": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Audience Fit"
          },
          "specificity": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Specificity"
          },
          "risk_penalty": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Risk Penalty"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "product_match",
          "intent_value",
          "evidence_strength",
          "audience_fit",
          "specificity",
          "risk_penalty"
        ],
        "title": "ScoreComponents"
      },
      "StageState": {
        "type": "string",
        "enum": [
          "pending",
          "succeeded",
          "warning",
          "failed"
        ],
        "title": "StageState"
      },
      "StageStatus": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "state": {
            "$ref": "#/components/schemas/StageState"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "state"
        ],
        "title": "StageStatus"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}